feat: fix loop
